Migrate database to Turso for improved scalability and performance
This commit refactors the database layer to use Turso (SQL) by updating configurations in `drizzle.config.ts`, replacing the PostgreSQL client with `@libsql/client` in `server/db.ts`, and adjusting schema definitions in `shared/schema.ts` from PostgreSQL to SQLite. It also includes necessary SQL syntax conversions in `server/episodic-memory.ts` and `server/guild-memory.ts` to accommodate SQLite, and gracefully stubs out pgvector functionality in `server/semantic-memory.ts` due to incompatibility. Additionally, a type cast in `server/gacha.ts` is corrected.

+++ server/gacha.ts
@@ -139,7 +139,7 @@ export async function getCollection(userId: string, page = 1, pageSize = 10): Pr
   const offset = (page - 1) * pageSize;
   const [cards, countResult] = await Promise.all([
     db.select().from(waifuCollection).where(eq(waifuCollection.userId, userId)).orderBy(desc(waifuCollection.rarityStars), desc(waifuCollection.obtainedAt)).limit(pageSize).offset(offset),
-    db.select({ count: sql<number>`count(*)::int` }).from(waifuCollection).where(eq(waifuCollection.userId, userId)),
+    db.select({ count: sql<number>`count(*)` }).from(waifuCollection).where(eq(waifuCollection.userId, userId)),
   ]);
   const total = countResult[0]?.count ?? 0;
   return { cards, total, page, totalPages: Math.max(1, Math.ceil(total / pageSize)) };
